remove ServiceEventCount compatibility test
Remove TestExecutionResultConsistency_ChunkServiceEventCountField.
As it relied on the internals of flow-go, this test would only be comparing
one reimplementation of the old ID hashing against another reimplementation,
for what I think is not much benefit.

TestVerifyExecutionResultHash and TestVerifyBlockHash can still provide
coverage for the ServiceEventCount field, provided the appropriate versions
and block heights.